FIX: SPS Relatred Refactoring and Fixed in Prompt Service (#550)
* Refactored the prompt service related to SPS

* Improved column names for manage documents modal

---------
